refactor(autonomi): internal refactors
